Network showing cosine similarity between clusters according to GSEA result

Network showing cosine similarity between clusters according to GSEA result

Cosine_networkByGSEA(
  SeuratObj,
  by = "GO",
  layout = layout_nicely,
  cos_sim_thresh = 0.6,
  p.adjust_thresh = 0.05,
  SEED = 123,
  node.cex = 5,
  width_range = c(0.8, 4),
  text_size = 1,
  vertex.label.dist = 0.5
)

Arguments

SeuratObj
Seurat object

by
GO KEGG Reactome MSigDb WikiPathways DO NCG DGN.

layout
layout_nicely, layout_with_fr, etc. Either a function or a numeric matrix. It specifies how the vertices will be placed on the plot.

cos_sim_thresh
only draw links of cosine similarity bigger than this

p.adjust_thresh
threshold of p.adjust to filter pathways used to calculate cosine similarity

SEED
seed

node.cex
node size

width_range
range of width of links

text_size
size of text

vertex.label.dist
The distance of the label from the center of the vertex.

Examples

if (FALSE) {
Cosine_networkByGSEA(SeuratObj, layout=layout_nicely, cos_sim_thresh=0.6,
                     p.adjust_thresh=0.05, SEED=123, node.cex=3, width_range=c(0.1, 0.8))
Cosine_networkByGSEA(SeuratObj, layout=layout_with_fr, cos_sim_thresh=0.8,
                     p.adjust_thresh=0.05, SEED=123, node.cex=5, width_range=c(0.8, 4))
}