Heatmap showing unique and shared pathways of clusters.

Heatmap showing unique and shared pathways of clusters.

pathway_unique_shared(
  SeuratObj,
  by = "GO",
  toshow = "-logFDR",
  topPath = 5,
  colour = "Greens",
  scale = "none",
  fontsize_row = 10,
  cluster_rows = TRUE,
  cluster_cols = TRUE
)

Arguments

SeuratObj

Seurat object

by

which GSEA result to show, one of "GO", "KEGG", "Reactome", "MSigDb", "WikiPathways", "DO", "NCG", "DGN"

toshow

which GSEA score to show, "-logFDR", "enrichmentScore", "NES", "pvalue", "p.adjust"

topPath

number of top pathways of each cluster to show

colour

color of heatmap, see RColorBrewer::brewer.pal.info

scale

if the values should be centered and scaled in either the row direction or the column direction, or none. Corresponding values are "row", "column" and "none"

fontsize_row

fontsize for rownames

cluster_rows

boolean values determining if rows should be clustered.

cluster_cols

boolean values determining if columns should be clustered.

Value

A Heatmap-class object.

Examples

if (FALSE) {
pathway_unique_shared(SeuratObj, by = "GO", toshow = "-logFDR", topPath = 5, colour = "Greens")
}